refactor: dissolve `actor` module into parent directory
