--- packages/backend/src/actor/mod.rs
@@ -1,2 +1,0 @@
-pub mod event;
-pub mod message;
